move session timeout to session config, use new permission handler
