+++ src/server/api/client/[clientId]/disable.post.ts
@@ -1,8 +1,11 @@
-export default defineEventHandler(async (event) => {
-  const { clientId } = await getValidatedRouterParams(
-    event,
-    validateZod(clientIdType)
-  );
-  await WireGuard.disableClient({ clientId });
-  return { success: true };
-});
+export default definePermissionEventHandler(
+  actions.CLIENT,
+  async ({ event }) => {
+    const { clientId } = await getValidatedRouterParams(
+      event,
+      validateZod(clientIdType)
+    );
+    await WireGuard.disableClient({ clientId });
+    return { success: true };
+  }
+);
