+++ src/server/api/client/[clientId]/index.get.ts
@@ -1,7 +1,10 @@
-export default defineEventHandler(async (event) => {
-  const { clientId } = await getValidatedRouterParams(
-    event,
-    validateZod(clientIdType)
-  );
-  return WireGuard.getClient({ clientId });
-});
+export default definePermissionEventHandler(
+  actions.CLIENT,
+  async ({ event }) => {
+    const { clientId } = await getValidatedRouterParams(
+      event,
+      validateZod(clientIdType)
+    );
+    return WireGuard.getClient({ clientId });
+  }
+);
