+++ src/server/api/client/[clientId]/index.post.ts
@@ -1,18 +1,21 @@
-export default defineEventHandler(async (event) => {
-  const { clientId } = await getValidatedRouterParams(
-    event,
-    validateZod(clientIdType)
-  );
+export default definePermissionEventHandler(
+  actions.CLIENT,
+  async ({ event }) => {
+    const { clientId } = await getValidatedRouterParams(
+      event,
+      validateZod(clientIdType)
+    );
 
-  const data = await readValidatedBody(
-    event,
-    validateZod(clientUpdateType, event)
-  );
+    const data = await readValidatedBody(
+      event,
+      validateZod(clientUpdateType, event)
+    );
 
-  await WireGuard.updateClient({
-    clientId,
-    client: data,
-  });
+    await WireGuard.updateClient({
+      clientId,
+      client: data,
+    });
 
-  return { success: true };
-});
+    return { success: true };
+  }
+);
