+++ src/server/api/client/[clientId]/qrcode.svg.get.ts
@@ -1,9 +1,12 @@
-export default defineEventHandler(async (event) => {
-  const { clientId } = await getValidatedRouterParams(
-    event,
-    validateZod(clientIdType)
-  );
-  const svg = await WireGuard.getClientQRCodeSVG({ clientId });
-  setHeader(event, 'Content-Type', 'image/svg+xml');
-  return svg;
-});
+export default definePermissionEventHandler(
+  actions.CLIENT,
+  async ({ event }) => {
+    const { clientId } = await getValidatedRouterParams(
+      event,
+      validateZod(clientIdType)
+    );
+    const svg = await WireGuard.getClientQRCodeSVG({ clientId });
+    setHeader(event, 'Content-Type', 'image/svg+xml');
+    return svg;
+  }
+);
