+++ src/server/api/client/index.get.ts
@@ -1,3 +1,3 @@
-export default defineEventHandler(() => {
+export default definePermissionEventHandler(actions.CLIENT, () => {
   return WireGuard.getClients();
 });
