+++ src/server/api/client/index.post.ts
@@ -1,11 +1,14 @@
 import { ClientCreateSchema } from '#db/repositories/client/types';
 
-export default defineEventHandler(async (event) => {
-  const { name, expiresAt } = await readValidatedBody(
-    event,
-    validateZod(ClientCreateSchema)
-  );
-  await Database.clients.create({ name, expiresAt });
-  await WireGuard.saveConfig();
-  return { success: true };
-});
+export default definePermissionEventHandler(
+  actions.CLIENT,
+  async ({ event }) => {
+    const { name, expiresAt } = await readValidatedBody(
+      event,
+      validateZod(ClientCreateSchema)
+    );
+    await Database.clients.create({ name, expiresAt });
+    await WireGuard.saveConfig();
+    return { success: true };
+  }
+);
