+++ src/server/api/wireguard/backup.get.ts
@@ -1,6 +1,9 @@
-export default defineEventHandler(async (event) => {
-  const config = await WireGuard.backupConfiguration();
-  setHeader(event, 'Content-Disposition', 'attachment; filename="wg0.json"');
-  setHeader(event, 'Content-Type', 'text/json');
-  return config;
-});
+export default definePermissionEventHandler(
+  actions.ADMIN,
+  async ({ event }) => {
+    const config = await WireGuard.backupConfiguration();
+    setHeader(event, 'Content-Disposition', 'attachment; filename="wg0.json"');
+    setHeader(event, 'Content-Type', 'text/json');
+    return config;
+  }
+);
