+++ src/server/api/wireguard/restore.put.ts
@@ -1,5 +1,8 @@
-export default defineEventHandler(async (event) => {
-  const { file } = await readValidatedBody(event, validateZod(fileType));
-  await WireGuard.restoreConfiguration(file);
-  return { success: true };
-});
+export default definePermissionEventHandler(
+  actions.ADMIN,
+  async ({ event }) => {
+    const { file } = await readValidatedBody(event, validateZod(fileType));
+    await WireGuard.restoreConfiguration(file);
+    return { success: true };
+  }
+);
