--- src/server/database/repositories/general/schema.ts
@@ -1,16 +1,0 @@
-import { sql } from 'drizzle-orm';
-import { int, sqliteTable, text } from 'drizzle-orm/sqlite-core';
-
-export const general = sqliteTable('general_table', {
-  // limit to one entry
-  id: int().primaryKey({ autoIncrement: false }).default(1),
-  //test: text().notNull(),
-  sessionTimeout: int('session_timeout').notNull(),
-  createdAt: text('created_at')
-    .notNull()
-    .default(sql`(CURRENT_TIMESTAMP)`),
-  updatedAt: text('updated_at')
-    .notNull()
-    .default(sql`(CURRENT_TIMESTAMP)`)
-    .$onUpdate(() => sql`(CURRENT_TIMESTAMP)`),
-});
