+++ src/server/database/repositories/sessionConfig/schema.ts
@@ -5,6 +5,7 @@ export const sessionConfig = sqliteTable('session_config_table', {
   // limit to one entry
   id: int().primaryKey({ autoIncrement: false }).default(1),
   password: text().notNull(),
+  timeout: int().notNull(),
   createdAt: text('created_at')
     .notNull()
     .default(sql`(CURRENT_TIMESTAMP)`),
